Moved logic for updating pupilStatus redis cache into a service
